code: изменение типа на double

+++ main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main()
 {
     cout << "Enter A and B: ";
-    int a, b;
+    double a, b;
     cin >> a >> b;
     cout << "A + B = " << a + b << '\n'
          << "A - B = " << a - b << '\n'
@@ -17,6 +17,7 @@ int main()
     }
     else
     {
+
         cout<<"max = "<<b<< '\n';
     }
     return 0;
